+++ app/encoding.go
@@ -3,6 +3,7 @@ package app
 import (
 	"github.com/sei-protocol/sei-chain/app/params"
 	retiredibcgov "github.com/sei-protocol/sei-chain/app/retiredibc/gov"
+	"github.com/sei-protocol/sei-chain/app/retiredoracle"
 	"github.com/sei-protocol/sei-chain/sei-cosmos/std"
 )
 
@@ -14,6 +15,8 @@ func MakeEncodingConfig() params.EncodingConfig {
 	ModuleBasics.RegisterLegacyAminoCodec(encodingConfig.Amino)
 	ModuleBasics.RegisterInterfaces(encodingConfig.InterfaceRegistry)
 	retiredibcgov.RegisterInterfaces(encodingConfig.InterfaceRegistry)
+	retiredoracle.RegisterInterfaces(encodingConfig.InterfaceRegistry)
+	retiredoracle.RegisterLegacyAminoCodec(encodingConfig.Amino)
 	return encodingConfig
 }
 
@@ -25,5 +28,7 @@ func MakeLegacyEncodingConfig() params.EncodingConfig {
 	ModuleBasics.RegisterLegacyAminoCodec(encodingConfig.Amino)
 	ModuleBasics.RegisterInterfaces(encodingConfig.InterfaceRegistry)
 	retiredibcgov.RegisterInterfaces(encodingConfig.InterfaceRegistry)
+	retiredoracle.RegisterInterfaces(encodingConfig.InterfaceRegistry)
+	retiredoracle.RegisterLegacyAminoCodec(encodingConfig.Amino)
 	return encodingConfig
 }